requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.3.2
colorama==0.4.6
fastapi==0.135.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.6
pyaes==1.6.1
pyasn1==0.6.3
pydantic==2.12.5
pydantic_core==2.41.5
Pyrogram==2.0.106
PySocks==1.7.1
python-multipart==0.0.24
python-telegram-bot==22.7
requests==2.31.0
rsa==4.9.1
starlette==1.0.0
Telethon==1.42.0
TgCrypto==1.2.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.15
uvicorn==0.44.0
